Added a temporary C++ workaround for __SMACK_value
I have declared __SMACK_value for C++ as a variadic function, with
C linkage to prevent name mangling.
The benefit of this declaration is that clang++ produces LLVM IR
code for this variadic function that closely resembles the IR code
that clang produces for the __SMACK_value() function. Hence, smack
can parse the IR code; no changes toSmackRep.cpp are necessary.

With this hack, I have successfully tested a C++ variant of
ct-verif's Tiny Encryption Algorithm example (see post below
for details).

The problem with this workaround is that the type that
__SMACK_value takes is now hard-coded into the declaration,
in other words, it currently only works for unsigned int*
(which I happened to need for the Tiny Encryption Algorithm example).

Related to issue:
  https://github.com/smackers/smack/issues/353